Update README.md
This project encompasses a web application that processes and visualizes data obtained from the publications section of a personal website. The primary goal of the project is to extract data from the web page using the Selenium module, create a backend utilizing the Flask framework, and develop a frontend using the Jinja2 template engine.